requirements.txt
pandas==2.1.4
numpy==1.26.3
scikit-learn==1.4.0
xgboost==2.0.3
mlflow==2.9.2
dvc==3.38.1
pyyaml==6.0.1
openpyxl==3.1.2
matplotlib==3.8.2
seaborn==0.13.1
joblib==1.3.2
